requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
langgraph>=0.0.30
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-anthropic>=0.1.5
langsmith>=0.0.80
httpx>=0.25.0
pytest>=7.4.3
python-multipart>=0.0.6 